Fix Perforce-related unit tests when Perforce isn't installed.
This makes the usage of `P4` conditional in the Perforce unit tests,
which is important for testing as it's becoming harder to use Perforce
on Python 2.7 in some newer environments due to OpenSSL compatibility
issues.

This is a backport of parts of commits a32b331849 and 6c9e34b64e.