Fix the branch determination for stored code when building the docs.
When building docs, we compute the release branch for the product, based
on the version. The code we had for this was using indexes into
`reviewboard.VERSION`, and had a wrong index used in one spot.

This was hit when building beta docs, as it's a code path we don't usually
hit during doc building.

This change bullet-proofs this logic by pulling out all the variants
from `VERSION` up-front and then uses those for the logic, instead of
index lookups.

Testing Done:
Successfully built the docs.

Reviewed at https://reviews.reviewboard.org/r/12472/