Fix unit test regressions with the latest Python-Markdown.
Our unit tests have been using the Python-Markdown version to determine
what results to expect when rendering. We were using an outdated
variable for this version, and that variable was just removed in this
week's Python-Markdown release.

This change updates our unit tests to check the modern variable. This is
compatible with all supported versions of Python-Markdown.

Testing Done:
Unit tests pass for Python 2.7 and 3 with the latest Python-Markdown on
each.

Reviewed at https://reviews.reviewboard.org/r/12477/