Limit data returned from quick-search API requests.
The search box in the upper right uses the search API for its
auto-complete. This was just doing the API request with the query and
then pulling out the information it needs, but the results from that API
request includes a whole lot of stuff that we really don't care about.
This change adds limiting to that API request so we only get the data
which is used to display the search results in the drop-down, along with
the URL.

Testing Done:
- Did a bunch of quick searches to verify that review requests, users,
  and groups were all displayed correctly and opened the correct URLs
  when clicked.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/12484/