Add release notes for Review Board 4.0.8.
This is a small bug fix and compatibility update. It caps the versions
of Python-Markdown on Python 3 and fixes running `rbext` on Python 3.10.

It alo fixes a bug with switching between Source/Rendered tabs in the
Markdown review UI and fixes a bug that could manifest in very new
installations where comments of different types with the same database
IDs on the same review request would all appear to resolve or reopen
when toggling the issue state for one of them.

Testing Done:
Built the docs. Checked for spelling errors, build errors, and bad links.

Reviewed at https://reviews.reviewboard.org/r/12486/