Add a management command for invalidating tokens
This change adds an `rb-site manage /path/to/site invalidate-api-tokens` command
that allows administrators to invalidate the API tokens for a set of users. Now
tokens can be invalidated individually via the API, or in batch with this
command.

Testing Done:
Manually tested the command and checked that the right tokens were
being invalidated.

Reviewed at https://reviews.reviewboard.org/r/12481/