Add release notes for Review Board 5.0 beta 2.
This release focuses on enhanced API tokens, repository ACL APIs (which
will be backported to 4.0), some performance improvements, and other bug
fixes.

THis change adds the release notes, future-proofs some Sphinx
configuration and references for 5.x releases, and tweaks the docs on
the `invalidate-api-tokens` management command to let us link and to
show the version in which this was added.

Testing Done:
Built the release notes. Checked for build errors, bad links, and spelling
errors.

Reviewed at https://reviews.reviewboard.org/r/12542/