Make it possible to skip the login form entirely when using SSO.
One requirement we've received from an early user of the new SSO
implementation is that they need to be able to entirely disable the
traditional login form. This change adds a new secret siteconfig key
which can be set to the name of an (active) SSO backend. When set, the
login page will redirect to the SSO login flow directly instead of
rendering the form.

This also affected the logout view, because we were using Django's
`logout_then_login` to handle this. The result of that is we would log
out, then immediately load the login view, which would then log the user
back in. I've changed it so if we're using the auto login functionality,
then the logout view will instead just use Django's plain logout, which
renders a template saying the user has logged out.

Testing Done:
- Set the key and saw that actions which would previously show the login
  form immediately began the SSO login flow (which could end up logging
  the user in with no action required depending on their session with
  the SSO provider). Logged out and saw it just go to a basic log out
  confirmation page.
- Ran unit tests.