Move and update search settings documentation.
The search documentation was still found deep under "Site Maintenance",
since it originally was just a small doc that explained how to set up
the crontab to run the indexing command. These days we have extensive
setup options (various Elasticsearch dependencies), and ideally if
people are setting up search, they'll also be using on-the-fly indexing.
This change moves the search docs to exist alongside the other documents
for the site settings, and renames it to "Search Settings" since it's
about a lot more than just indexing. While I was doing this, I also
moved the Elasticsearch section to be first, since we definitely want
users to prioritize that over Whoosh.

Testing Done:
- Built HTML docs. Verified that all the links went to the expected
  places, and that the search doc rendered as expected.

Reviewed at https://reviews.reviewboard.org/r/12585/