Fix API token not being displayed after creation.
When creating an API token through the API tokens config page, the page would
need to be reloaded in order for the token to show up in the list of tokens.
This was due to the token's `expired` field being missing from the list of
attributes for that token item, which prevented the token item from
rendering. This change fixes that, and corrects a separate minor issue in the
docs.

Testing Done:
Created new API tokens from the config page and made sure they appeared
immediately on the page.

Reviewed at https://reviews.reviewboard.org/r/12583/