Fix incorrect profile_columns_field for the reviews datagrid.
The reviews datagrid has had the `profile_columns_field` set to a
nonexistent field name for a long time, but a recent change in Djblets
exposed this as an error. We don't particularly care about this, so we
can just remove it.

Testing Done:
Accessed the reviews datagrid and saw the rendered page instead of an
error.

Reviewed at https://reviews.reviewboard.org/r/12598/