Add stale Ship It! styling and ARIA support to the Ship It!/Issues column.
This updates the column to help identify when a review request has been
updated since receiving one or more Ship It! reviews. Now, if a review
request has been updated, the Ship It! count still remains, but the
color is grey instead of green, making it stand out and helping
reviewers know that they should look at the change again.

This also makes this column accessible. The indicator has an ARIA label
and title that convey the contents, with the icon and count being hidden
from the accessibility tree.

The text is localized based on the number of items, so there's a
different string for 1 Ship It or issue vs. many.

Testing Done:
Tested with every combination of Ship It/Issue state and stale state,
verifying the tooltip and presentation.

Unit tests pass.

Bugs closed: 2106

Reviewed at https://reviews.reviewboard.org/r/12593/