Move TokenExpiresFieldType from Djblets into Review Board.
Djblets gained a `TokenExpiresFieldType`, which was a
`DateTimeFieldType` subclass that allowed for empty strings.

We'll eventually want to allow fields like `DateTimeFieldType` to
optionally take an empty value, but we'll want to do that through an
official mechanism that all field types can opt into. Since this
particular class is very specific to `APITokenResource`, this change
moves it into that module, making it effectively a private
implementation detail.

The field type will be removed from Djblets, so we don't have to worry
about supporting it and then deprecating it down the road.

Testing Done:
API token resource unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/12614/