Fix a missing column of background color in code blocks.
The code block CSS had a vertical strip of white just to the left of the
actual code being displayed. This was due to a `:before` rule that
didn't have a background set, getting the background instead from the
parent container.

This change adds a `background: inherit` to ensure it gets the
background color set by the element it's part of.

Testing Done:
Tested that this looked correct on Firefox and Chrome.

Checked code blocks in text fields (like Description), which dock to edges,
and in reviews.

Reviewed at https://reviews.reviewboard.org/r/12615/