Cap the version of Django Evolution for Review Board 4.0.x.
We will soon be releasing Django Evolution 2.2, and while this should be
safely backwards-compatible with 2.1 (the minimum version series
required by Review Board 4.0.x), we want to aim for stability at this
point.

As a precaution, we now cap Django Evolution to 2.1.x for this release.

Testing Done:
Installed the package and verified it got a 2.1.x release.

Reviewed at https://reviews.reviewboard.org/r/12576/