Include the Python version and install location in rbssh --version.
`rbssh --version` now includes the same information found when running
`rb-site --version`. This includes the Python version and the location
in which `rbssh` was installed.

This can help with diagnosing issues where there may be multiple
versions of `rbssh` potentially installed (such as when a user attempts
to upgrade and switches Python versions in the process, but the old copy
remains).

Testing Done:
Ran with `--version` and saw the output.

Reviewed at https://reviews.reviewboard.org/r/12575/