Fix a crash in rb-site when the pseudo terminal returns a width of 0.
When running on a pseudo-terminal (such as in a NAS),
`get_terminal_size()` could return (0,0).

When this happens, rb-site ends up calculating widths that end up
negative, since it tries to keep padding on the side of the terminal.
This ultimately crashes.

Testing Done:
By hand, on a Synology DiskStation DS220+, with RB 4.0.7 running in a
Docker container.

Reviewed at https://reviews.reviewboard.org/r/12464/