Fix multi-commit uploading with Mercurial in pre-commit review mode.
Uploading multi-commit review request diffs on Mercurial fails with File
Not Found errors. This only fails in multi-commit mode, and only when
different files are changed in different diffs.

There's discussion on this topic here:
https://groups.google.com/g/reviewboard/c/FG5TNTNC_fU

This fails due to a lack of a base commit ID. It tries to request a raw
file from a server repository using the revision ID of the previous
uploaded commit, but there's no such revision on the server as those
patches aren't pushed yet to the upstream repository in a pre-commit
review case.

Passing a base commit ID allows this to use the correct revision to find
these files.

Testing Done:
Reproduced the original problem in a new database, and verified that
this fixes the issue.

Reviewed at https://reviews.reviewboard.org/r/12623/