Add support for a User.FullName SAML attribute
If User.FirstName or User.LastName is not specified, we try to look for
a User.FullName attribute and split that on the first space to get a
first and last name.

This was necessary because some IdPs only let you specify a full name.

Testing Done:
Applied this in our container build and tested it live at
reviews.imfreedom.org using JetBrains Hub as the IdP.

Reviewed at https://reviews.reviewboard.org/r/12633/