Consolidate and clarify API token documentation.
The documentation for creating API tokens was split into two different
places, and we had different content for each. This change moves
everything into one place and links from the other. It also clarifies
the behavior for token expiration a bit to match the current reality.

I was initially thinking of a screenshot in here, but given the
surrounding content, I don't know that we need it.

Testing Done:
Built HTML and checked output.

Reviewed at https://reviews.reviewboard.org/r/12655/