Add release notes for Review Board 5.0.
This is a large release, moving onto modern architecture, and
introducing many new features for reviewers and administrators. This
includes Stale Ship-It! indicators, Trojan Source detection, Single
Sign-On, and more.

This change adds release notes covering all the public changes we want
to advertise for the release.

This lacks screenshots, but the plan is to add them in a follow-up
change.

Testing Done:
Built the release notes. Checked for spelling errors, bad links, and
build errors.

Reviewed at https://reviews.reviewboard.org/r/12660/