Split EmailTestHelper out from test_email_sending.py
We have a helper class that handles much of the setup required to test
that reviews and review requests send emails, but this was contained
within the test case file where it was used. I need to use it elsewhere,
so I've split it out into a new file to import from.

Testing Done:
Used the helper from a new test file.

Reviewed at https://reviews.reviewboard.org/r/12693/