Fix reverting to wrong policy when canceling out of the API token custom policy editor.
When canceling out of the API token custom policy editor, the policy would
revert to the 2nd previous value instead of the most previous value. This made
it so that sometimes the policy would revert to the wrong value. This change
fixes that so that the policy reverts to the most previous value when
canceling out of the custom policy editor.

Testing Done:
- Manually tested canceling out of the policy editor and made sure
  the right policy was reverted to.

Reviewed at https://reviews.reviewboard.org/r/12668/