Add a new batch operation view
For upcoming feature work, we want the ability to make a single request
to publish a bunch of individual items (review requests and reviews) all
together.

To do this, we're introducing a new API-like endpoint for performing
batch operations. This doesn't exist in the official API because it
doesn't really mesh well with REST, and the actual endpoint may be in
flux for a bit.

We already have some batch-like operations in the dashboard UI that
currently just make a whole bunch of API requests for closing review
requests or changing visibility state. This new endpoint will also
enable these workflows, so we can handle it much more efficiently.

The new batch endpoint supports these operations:
* close
* discard
* archive
* mute
* unarchive (effectively unmute too)
* publish

Testing Done:
- Looked at e-mail previews for batch publishes with all combinations of
  review requests, review request updates, reviews, and review replies in
  both HTML and text mode.
- Used all operations except for publish from real UI.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12695/