Fix up arguments to logging methods.
We have a special logging formatter that is used to show info when a
django `HttpRequest` is part of the log record. Once upon a time, the
way of getting extra information into a log record was just to pass it
as additional `kwargs` to the log methods, but the modern, documented
way of doing it is to pass a single `extra` dict. This change fixes our
code up to use that. We also had several logging instances where we had
the request available but were not passing it to the log.

While I was auditing our log messages, I fixed up some other trivial
issues like formatting with args and indentation.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12705/