Display deprecation notices on the API tokens config page and deprecation status in the resource.
This change adds a `deprecated` field to the API token web API resource. This
is a bool showing whether the token was generated by a deprecated token
generator. We also now display a notice on the API tokens config page if a
token is deprecated. The notice encourages the user to delete the deprecated
token and generate a new one.

Testing Done:
- Made some GET requests to the `/api-tokens/` list endpoint and saw that the
  `deprecated` field was included in the token resources.
- Created some deprecated tokens and made sure they behave correctly and
  appear with the notice in the API tokens page.
- Ran JS unit tests.
- Ran `reviewboard/webapi/tests/test_api_token.py`

Reviewed at https://reviews.reviewboard.org/r/12631/