Add change description return to ReviewRequest.publish.
The draft publish operation returns a change description, but this
wasn't plumbed through for users of `ReviewRequest.publish`. This is
potentially useful in many situations, and necessary for my work with
the batch publish operation.

Testing Done:
Used this in conjunction with other changes.

Reviewed at https://reviews.reviewboard.org/r/12714/