Add author name as an attribute to RB.Review in JS resources.
In order to make it so we can show "Replying to X's review" in the new
unified banner, we need a reasonable way of getting the author name for
the review. This change adds it as an attribute for the JS model, and
populates it as part of the entry data.

Testing Done:
Used in conjunction with my WIP unified banner change.

Reviewed at https://reviews.reviewboard.org/r/12733/