Change default dashboard view to "Overview".
We added the overview mode to the dashboard a while ago, and it really
seems like it's what most people want--a list of both their incoming and
outgoing review requests. Unfortunately, it took extra clicks to get to
it every single time. This change makes it the default view for the
dashboard.

Testing Done:
Opened the dashboard and saw that it went straight to the overview view.

Reviewed at https://reviews.reviewboard.org/r/12728/