Support additional mimetypes for text-based file attachments.
This adds support for various alternative JSON mimetypes and YAML
mimetypes, ensuring that thumbnails display content correctly and that
they can be reviewed as text files when using an `application/`
mimetype.

The text-based review UI uses the text-based mimetype as the basis for
the list of supported mimetypes, so these won't get out of sync.

Testing Done:
Unit tests pass.

Manually tested with the new mimetypes, ensuring that thumbnails and
reviews worked correctly.

Verified the fix with a customer.

Reviewed at https://reviews.reviewboard.org/r/12730/