Add bulletproofing to the bug views.
We're hitting some crashes on the demo server from bots accessing the
bug redirect view for a review request that does not have an associated
repository. This view made an assumption that the repository exists,
which is generally fine for normal usage but fails when crawlers try to
be clever.

This change adds some bulletproofing to the bug-related views to
validate that the repository exists and return a 404 if it doesn't.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12751/