Add base classes and registry for new actions framework.
The actions framework is in the process of being rewritten to give us
better consistency across all the different types of actions, a real
registry, and new features and capabilities.

This change introduces the actions app with the action base classes and
registry, along with a JavaScript-side stub model/view, template tags
for rendering the actions on pages, and the beginnings of some unit
tests.

Testing Done:
- Ran unit tests.
- Used this with other changes that start porting existing actions over
  to the new framework.

Reviewed at https://reviews.reviewboard.org/r/12752/