Bullet-proof bugs closed field to prevent newlines.
Copy/pasting values into the bugs field can sometimes save values that
include newline characters. These aren't visible to the user, and don't
have too many ill effects, but it was discovered that attempting to use
the API to download the diff would cause a crash when the bugs were used
to indicate the filename for the download.

This change adds two-stages of bullet-proofing. The field itself has
been changed to filter out newline characters when saving, to prevent
this from occurring in the future. For databases that already include
newlines in this field, the diff API resource has been updated to
prevent errors when creating the headers for the diff download.

Testing Done:
- Copy/pasted values with newlines into the bugs field with various
  browsers and saw that the saved values were correct.
- Ran unit tests.